requirements.txt
aiohttp
python-dotenv
motor
slack-bolt
uvloop